make "#download" page work without js
The panel-navigation introduced in the prev commit, which works without javascript,
(using css), has a restriction: all panels/pages have to be contained in a
single html file. To accommodate this, this commit inlines panel-download.html
into index.html.

Previously we had
- panel-download.html.template => panel-download.html
- index.html

Now we have
- panel-download.html.template + index.html.template => index.html

electrum/contrib/make_download in the client repo is also adapted.